backend/app/core/security.py
"""Security utilities for password hashing and JWT tokens."""

import bcrypt
from jose import jwt
from datetime import datetime, timedelta, timezone

from app.core.config import settings


def verify_password(plain: str, hashed: str) -> bool:
    """Verify a plain password against a hashed password."""
    # bcrypt.checkpw expects bytes
    return bcrypt.checkpw(plain.encode("utf-8"), hashed.encode("utf-8"))


def hash_password(password: str) -> str:
    """Hash a password using bcrypt."""
    # bcrypt expects bytes and returns bytes
    pwd_bytes = password.encode("utf-8")
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(pwd_bytes, salt)
    return hashed.decode("utf-8")


def create_access_token(data: dict, expires_delta: timedelta | None = None) -> str:
    """Create a JWT access token."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm="HS256")
    return encoded_jwt
